api_for_llm/requirements.txt
fastapi==0.109.2
uvicorn==0.27.0.post1
ollama==0.1.5
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.6.1
typing-extensions==4.9.0
